- Updated SMS Communications so they don't appear to succeed (but really don't) when the From number is not specified - Updated CommunicationEntryWizard to check that the SMS From number was specified - Updated CommunicationEntryWizard to include extra logic to ensure that 'Test Communication' database artifacts get cleaned up

+++ RockWeb/Blocks/Communication/CommunicationEntryWizard.ascx.cs
@@ -1236,6 +1236,15 @@ namespace RockWeb.Blocks.Communication
                 return;
             }
 
+            try
+            {
+                CleanupOrphanedTestCommunications();
+            }
+            catch( Exception ex)
+            {
+                this.LogException( ex );
+            }
+
             Rock.Model.Communication communication = UpdateCommunication( new RockContext() );
 
             if ( communication != null )
@@ -1244,9 +1253,12 @@ namespace RockWeb.Blocks.Communication
                 {
                     // Using a new context (so that changes in the UpdateCommunication() are not persisted )
                     int? testPersonId = null;
+                    Rock.Model.Communication testCommunication = null;
+                    CommunicationService communicationService = null;
+
                     try
                     {
-                        var testCommunication = communication.Clone( false );
+                        testCommunication = communication.Clone( false );
                         testCommunication.Id = 0;
                         testCommunication.Guid = Guid.Empty;
                         testCommunication.EnabledLavaCommands = GetAttributeValue( "EnabledLavaCommands" );
@@ -1296,7 +1308,10 @@ namespace RockWeb.Blocks.Communication
 
                         testPerson.ForeignGuid = null;
                         testPerson.ForeignId = null;
-                        testPerson.ForeignKey = null;
+
+                        // Just in case it doesn't get cleaned up, mark it so that we can try to clean it up next time
+                        testPerson.ForeignKey = "_ForTestCommunication_";
+
                         var personService = new PersonService( rockContext );
                         personService.Add( testPerson );
                         rockContext.SaveChanges();
@@ -1312,7 +1327,7 @@ namespace RockWeb.Blocks.Communication
 
                         testCommunication.Recipients.Add( testRecipient );
 
-                        var communicationService = new CommunicationService( rockContext );
+                        communicationService = new CommunicationService( rockContext );
                         communicationService.Add( testCommunication );
                         rockContext.SaveChanges();
 
@@ -1339,26 +1354,66 @@ namespace RockWeb.Blocks.Communication
 
                         nbResult.Dismissable = true;
                         nbResult.Visible = true;
-
-                        communicationService.Delete( testCommunication );
-                        rockContext.SaveChanges();
                     }
                     finally
                     {
-                        // make sure we delete the Person record we created to send the test email
+                        // make sure we delete the test communication record we created to send the test 
+                        if ( communicationService != null && testCommunication != null )
+                        {
+                            communicationService.Delete( testCommunication );
+                            rockContext.SaveChanges();
+                        }
+
+                        // make sure we delete the Person record we created to send the test 
                         using ( var deleteRockContext = new RockContext() )
                         {
+                            var personToDeleteService = new PersonService( deleteRockContext );
+                            var personAliasToDeleteService = new PersonAliasService( deleteRockContext );
+
                             if ( testPersonId.HasValue )
                             {
-                                var personToDeleteService = new PersonService( deleteRockContext );
-                                var personAliasToDeleteService = new PersonAliasService( deleteRockContext );
                                 var personToDelete = personToDeleteService.Get( testPersonId.Value );
                                 personAliasToDeleteService.DeleteRange( personToDelete.Aliases );
                                 personToDelete.Aliases = null;
                                 personToDeleteService.Delete( personToDelete );
                                 deleteRockContext.SaveChanges();
                             }
+}
+                    }
+                }
+            }
+        }
+
+        /// <summary>
+        /// Shouldn't happen, but just in case...
+        /// this will cleanup any orphaned test communications artifacts
+        /// </summary>
+        private void CleanupOrphanedTestCommunications()
+        {
+            using ( RockContext deleteRockContext = new RockContext() )
+            {
+                PersonService personToDeleteService = new PersonService( deleteRockContext );
+                PersonAliasService personAliasToDeleteService = new PersonAliasService( deleteRockContext );
+                
+                // check for any test communication artifacts that didn't clean up correctly
+                var forTestCommunicationPersonQry = personToDeleteService.Queryable().Where( a => a.ForeignKey == "_ForTestCommunication_" );
+                if ( forTestCommunicationPersonQry.Any() )
+                {
+                    var communicationToDeleteService = new Rock.Model.CommunicationService( deleteRockContext );
+                    var communicationToDeleteQry = communicationToDeleteService.Queryable().Where( a => a.Recipients.Any( r => forTestCommunicationPersonQry.Any( p => p.Id == r.PersonAlias.PersonId ) ) );
+                    foreach ( var testCommunicationToDelete in communicationToDeleteQry.ToList() )
+                    {
+                        communicationToDeleteService.Delete( testCommunicationToDelete );
+                    }
+
+                    foreach ( var personToDelete in forTestCommunicationPersonQry )
+                    {
+                        foreach ( var personToDeleteAlias in personToDelete.Aliases.ToList() )
+                        {
+                            personAliasToDeleteService.Delete( personToDeleteAlias );
                         }
+
+                        personToDeleteService.Delete( personToDelete );
                     }
                 }
             }
@@ -1813,6 +1868,15 @@ namespace RockWeb.Blocks.Communication
         /// <param name="e">The <see cref="EventArgs"/> instance containing the event data.</param>
         protected void btnSMSSendTest_Click( object sender, EventArgs e )
         {
+            if (ddlSMSFrom.SelectedValue.IsNullOrWhiteSpace())
+            {
+                nbSMSTestResult.NotificationBoxType = NotificationBoxType.Danger;
+                nbSMSTestResult.Text = "A 'From' number must be specified.";
+                nbSMSTestResult.Dismissable = true;
+                nbSMSTestResult.Visible = true;
+                return;
+            }
+
             SendTestCommunication( EntityTypeCache.Read( Rock.SystemGuid.EntityType.COMMUNICATION_MEDIUM_SMS.AsGuid() ).Id, nbSMSTestResult );
         }
